Refactor QuadTree implementation: replace old files with a new structure, improve node handling with concurrent vector, and enhance insertion logic for better performance
